+++ lib/wanikani/srs.rb
@@ -16,5 +16,24 @@ module Wanikani
       return srs_distribution if item_type == "all"
       return srs_distribution[item_type]
     end
+
+    # Gets all items for a specific SRS level.
+    #
+    # @param [String] item_type the SRS level for the items returned.
+    # @return [Array] Returns all the items matching the specified SRS for each level for the user.
+    def self.items_by_type(item_type)
+      raise ArgumentError, "Please use a valid SRS type." if !ITEM_TYPES.include?(item_type) || item_type == "all"
+
+      items_by_type = []
+      %w(radicals kanji vocabulary).each do |type|
+        items = Wanikani::Level.send(type)
+        items.reject! { |item| item["user_specific"].nil? || item["user_specific"]["srs"] != item_type }.map! do |item|
+          item.merge!("type" => (type == 'radicals' ? 'radical' : type))
+        end
+        items_by_type << items
+      end
+
+      items_by_type.flatten
+    end
   end
 end