+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # WaniKani gem - Changelog
 
+### 0.0.10
+
+- Added `Wanikani::SRS.items_by_type` to return all items for a specific SRS type.
+
 ### 0.0.9
 
 - Added an optional parameter to `Wanikani.valid_api_key?` to allow checking an API key without having to set `Wanikani.api_key` beforehand.